Move testCpCmd to helpers since it is used by both functional tests and multi node test.
